sort by length if filtered, dry
